% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DataGrid}
\alias{DataGrid}
\title{DataGrid}
\usage{
DataGrid(rows = NULL, columns = NULL, ...)
}
\arguments{
\item{rows}{rows}

\item{columns}{columns}

\item{...}{Other props}
}
\description{
DataGrid
}
