% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muiDataGridDependency.R
\name{muiDataGridDependency}
\alias{muiDataGridDependency}
\title{Mui X Data Grid UI JS dependency}
\usage{
muiDataGridDependency()
}
\value{
HTML dependency object.
}
\description{
Mui X Data Grid UI JS dependency
}
